Fixed a bug where an error would be thrown if the temporary directory was deleted when it tried to delete it.

+++ WebPebbleWQemu/Service/Containers/QemuController.cs
@@ -54,7 +54,8 @@ namespace WebPebbleWQemu.Service.Containers
             this.sessionDir = $"{Program.config.qemu_session_dir}sid_{sessionId.ToString()}/";
 
             //Ensure the session dir is new.
-            Directory.Delete(sessionDir, true);
+            if(Directory.Exists(sessionDir))
+                Directory.Delete(sessionDir, true);
             Directory.CreateDirectory(sessionDir);
 
             //Configure ports
